refactor(client): split Toolbar.tsx and SettingsPopover.tsx (#402)
Extract 5 focused sub-components and 1 shared style file; parent files
become thin coordinators with no behavior change.
