+++ src/client/components/settingsStyles.ts
@@ -1,0 +1,10 @@
+import type React from "react";
+
+export const sectionLabelStyle: React.CSSProperties = {
+  fontSize: "11px",
+  fontWeight: 600,
+  color: "var(--tandem-fg)",
+  marginBottom: "6px",
+  textTransform: "uppercase",
+  letterSpacing: "0.5px",
+};
